requirements.txt
requests==2.22.0
facebook-sdk==3.1.0
twitter==1.18.0
